Add my solution to exercise 4, which writes a parser for a simple external domain language.
Not all the commands of the language are
implemented, but that is easy to add and doesn't add much benefit.
There is some refactoring I could do with the handler functions, so that
there is only one function needed for the draw commands. The solution in
the back of the book shows how this could be done.